Extract MarketsTable component.
