+++ src/components/Marketing/index.js
@@ -20,6 +20,7 @@ import Hamburger from './Hamburger';
 import { BorrowCalculator, SaveCalculator } from './Calculators';
 import PageHead from './PageHead';
 import TokenIcon from './TokenIcon';
+import MarketsTable from './MarketsTable';
 
 import PageContentLayout from 'layouts/PageContentLayout';
 
@@ -75,5 +76,6 @@ export {
   SaveCalculator,
   StyledPageContentLayout,
   PageHead,
-  TokenIcon
+  TokenIcon,
+  MarketsTable
 };
